requirements.txt
django
djangorestframework
django-rest-auth
djangorestframework-jwt
pillow